herokuflashcard/requirements/test.txt
# requirements/test.txt
-r common.txt
django-debug-toolbar==0.9.4
